requirements.txt
aiohttp==3.7.4.post0
aioice==0.7.5
aiortc==1.2.0
async-timeout==3.0.1
attrs==21.2.0
av==8.0.3
cffi==1.14.5
chardet==4.0.0
crc32c==2.2
cryptography==3.4.7
dnspython==2.1.0
fast-colorthief==0.0.4
idna==3.2
ImageHash==4.2.0
multidict==5.1.0
netifaces==0.11.0
numpy==1.20.3
opencv-python==4.5.2.54
pandas==1.2.4
Pillow==8.2.0
pycparser==2.20
pyee==8.1.0
pylibsrtp==0.6.8
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
scipy==1.6.3
six==1.16.0
typing-extensions==3.10.0.0
yarl==1.6.3
